requirements.txt
mysql-connector-python>=9.6.0
openpyxl>=3.1.5
pandas>=3.0.0
google-adk
sqlglot
python-dotenv>=1.0.0
rich>=13.0.0
